implement labeled loops and next/redo/last LABEL for nqp itself on moar
By attaching a label to a loop three additional exception handlers get
installed: One for each of next, redo and last. Only the exception
handler that has the identical label (by object identity) will be in
charge for the exception. This does made labels only available for nqp
itself right now, but it should be a pattern for implementing it in P6.